StudentActivityDelteTest
