+++ src/test/java/acme/testing/student/enrolment/StudentEnrolmentDeleteTest.java
@@ -1,142 +1,142 @@
-
-package acme.testing.student.enrolment;
-
-import java.util.Collection;
-
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvFileSource;
-import org.springframework.beans.factory.annotation.Autowired;
-
-import acme.entities.Enrolment;
-import acme.testing.TestHarness;
-
-public class StudentEnrolmentDeleteTest extends TestHarness {
-
-	@Autowired
-	protected StudentEnrolmentTestRepository repository;
-
-
-	@ParameterizedTest
-	@CsvFileSource(resources = "/student/enrolment/delete-positive.csv", encoding = "utf-8", numLinesToSkip = 1)
-	public void test100Positive(final int recordIndex) {
-
-		super.signIn("student1", "student1");
-
-		super.clickOnMenu("Student", "My enrolments");
-		super.checkListingExists();
-		super.clickOnListingRecord(recordIndex);
-		super.clickOnSubmit("Delete");
-		super.checkNotErrorsExist();
-		super.checkListingExists();
-		super.sortListing(0, "asc");
-
-		super.signOut();
-	}
-
-	@ParameterizedTest
-	@CsvFileSource(resources = "/student/enrolment/create-negative.csv", encoding = "utf-8", numLinesToSkip = 1)
-	public void test200Negative(final int recordIndex, final String code, final String motivation, final String goals, final String workbook, final String course, final String creditCard) {
-
-	}
-
-	@Test
-	public void test300Hacking() {
-		// HINT: this test tries to delete an enrolment using principals with inappropriate roles
-
-		Collection<Enrolment> enrolments;
-		String param;
-
-		enrolments = this.repository.findManyEnrolmentsByStudentUsername("student1");
-		for (final Enrolment enrolment : enrolments) {
-			param = String.format("id=%d", enrolment.getId());
-
-			super.checkLinkExists("Sign in");
-			super.request("/student/enrolment/delete", param);
-			super.checkPanicExists();
-
-			super.signIn("administrator", "administrator");
-			super.request("/student/enrolment/delete", param);
-			super.checkPanicExists();
-			super.signOut();
-
-			super.signIn("student2", "student2");
-			super.request("/student/enrolment/delete", param);
-			super.checkPanicExists();
-			super.signOut();
-
-			super.signIn("assistant1", "assistant1");
-			super.request("/student/enrolment/delete", param);
-			super.checkPanicExists();
-			super.signOut();
-
-			super.signIn("company1", "company1");
-			super.request("/student/enrolment/delete", param);
-			super.checkPanicExists();
-			super.signOut();
-
-			super.signIn("lecturer1", "lecturer1");
-			super.request("/student/enrolment/delete", param);
-			super.checkPanicExists();
-			super.signOut();
-
-			super.signIn("auditor1", "auditor1");
-			super.request("/student/enrolment/delete", param);
-			super.checkPanicExists();
-			super.signOut();
-		}
-	}
-
-	@Test
-	public void test301Hacking() {
-		// HINT: this test tries to delete an enrolment with draftMode false
-
-		Collection<Enrolment> enrolments;
-		String param;
-
-		enrolments = this.repository.findManyEnrolmentsByStudentUsername("student1");
-		for (final Enrolment enrolment : enrolments)
-			if (!enrolment.isDraftMode()) {
-				param = String.format("id=%d", enrolment.getId());
-
-				super.checkLinkExists("Sign in");
-				super.request("/student/enrolment/delete", param);
-				super.checkPanicExists();
-
-				super.signIn("administrator", "administrator");
-				super.request("/student/enrolment/delete", param);
-				super.checkPanicExists();
-				super.signOut();
-
-				super.signIn("student1", "student1");
-				super.request("/student/enrolment/delete", param);
-				super.checkPanicExists();
-				super.signOut();
-
-				super.signIn("student2", "student2");
-				super.request("/student/enrolment/delete", param);
-				super.checkPanicExists();
-				super.signOut();
-
-				super.signIn("assistant1", "assistant1");
-				super.request("/student/enrolment/delete", param);
-				super.checkPanicExists();
-				super.signOut();
-
-				super.signIn("company1", "company1");
-				super.request("/student/enrolment/delete", param);
-				super.checkPanicExists();
-				super.signOut();
-
-				super.signIn("lecturer1", "lecturer1");
-				super.request("/student/enrolment/delete", param);
-				super.checkPanicExists();
-				super.signOut();
-
-				super.signIn("auditor1", "auditor1");
-				super.request("/student/enrolment/delete", param);
-				super.checkPanicExists();
-				super.signOut();
-			}
-	}
-}
+//
+//package acme.testing.student.enrolment;
+//
+//import java.util.Collection;
+//
+//import org.junit.jupiter.api.Test;
+//import org.junit.jupiter.params.ParameterizedTest;
+//import org.junit.jupiter.params.provider.CsvFileSource;
+//import org.springframework.beans.factory.annotation.Autowired;
+//
+//import acme.entities.Enrolment;
+//import acme.testing.TestHarness;
+//
+//public class StudentEnrolmentDeleteTest extends TestHarness {
+//
+//	@Autowired
+//	protected StudentEnrolmentTestRepository repository;
+//
+//
+//	@ParameterizedTest
+//	@CsvFileSource(resources = "/student/enrolment/delete-positive.csv", encoding = "utf-8", numLinesToSkip = 1)
+//	public void test100Positive(final int recordIndex) {
+//
+//		super.signIn("student1", "student1");
+//
+//		super.clickOnMenu("Student", "My enrolments");
+//		super.checkListingExists();
+//		super.clickOnListingRecord(recordIndex);
+//		super.clickOnSubmit("Delete");
+//		super.checkNotErrorsExist();
+//		super.checkListingExists();
+//		super.sortListing(0, "asc");
+//
+//		super.signOut();
+//	}
+//
+//	@ParameterizedTest
+//	@CsvFileSource(resources = "/student/enrolment/create-negative.csv", encoding = "utf-8", numLinesToSkip = 1)
+//	public void test200Negative(final int recordIndex, final String code, final String motivation, final String goals, final String workbook, final String course, final String creditCard) {
+//
+//	}
+//
+//	@Test
+//	public void test300Hacking() {
+//		// HINT: this test tries to delete an enrolment using principals with inappropriate roles
+//
+//		Collection<Enrolment> enrolments;
+//		String param;
+//
+//		enrolments = this.repository.findManyEnrolmentsByStudentUsername("student1");
+//		for (final Enrolment enrolment : enrolments) {
+//			param = String.format("id=%d", enrolment.getId());
+//
+//			super.checkLinkExists("Sign in");
+//			super.request("/student/enrolment/delete", param);
+//			super.checkPanicExists();
+//
+//			super.signIn("administrator", "administrator");
+//			super.request("/student/enrolment/delete", param);
+//			super.checkPanicExists();
+//			super.signOut();
+//
+//			super.signIn("student2", "student2");
+//			super.request("/student/enrolment/delete", param);
+//			super.checkPanicExists();
+//			super.signOut();
+//
+//			super.signIn("assistant1", "assistant1");
+//			super.request("/student/enrolment/delete", param);
+//			super.checkPanicExists();
+//			super.signOut();
+//
+//			super.signIn("company1", "company1");
+//			super.request("/student/enrolment/delete", param);
+//			super.checkPanicExists();
+//			super.signOut();
+//
+//			super.signIn("lecturer1", "lecturer1");
+//			super.request("/student/enrolment/delete", param);
+//			super.checkPanicExists();
+//			super.signOut();
+//
+//			super.signIn("auditor1", "auditor1");
+//			super.request("/student/enrolment/delete", param);
+//			super.checkPanicExists();
+//			super.signOut();
+//		}
+//	}
+//
+//	@Test
+//	public void test301Hacking() {
+//		// HINT: this test tries to delete an enrolment with draftMode false
+//
+//		Collection<Enrolment> enrolments;
+//		String param;
+//
+//		enrolments = this.repository.findManyEnrolmentsByStudentUsername("student1");
+//		for (final Enrolment enrolment : enrolments)
+//			if (!enrolment.isDraftMode()) {
+//				param = String.format("id=%d", enrolment.getId());
+//
+//				super.checkLinkExists("Sign in");
+//				super.request("/student/enrolment/delete", param);
+//				super.checkPanicExists();
+//
+//				super.signIn("administrator", "administrator");
+//				super.request("/student/enrolment/delete", param);
+//				super.checkPanicExists();
+//				super.signOut();
+//
+//				super.signIn("student1", "student1");
+//				super.request("/student/enrolment/delete", param);
+//				super.checkPanicExists();
+//				super.signOut();
+//
+//				super.signIn("student2", "student2");
+//				super.request("/student/enrolment/delete", param);
+//				super.checkPanicExists();
+//				super.signOut();
+//
+//				super.signIn("assistant1", "assistant1");
+//				super.request("/student/enrolment/delete", param);
+//				super.checkPanicExists();
+//				super.signOut();
+//
+//				super.signIn("company1", "company1");
+//				super.request("/student/enrolment/delete", param);
+//				super.checkPanicExists();
+//				super.signOut();
+//
+//				super.signIn("lecturer1", "lecturer1");
+//				super.request("/student/enrolment/delete", param);
+//				super.checkPanicExists();
+//				super.signOut();
+//
+//				super.signIn("auditor1", "auditor1");
+//				super.request("/student/enrolment/delete", param);
+//				super.checkPanicExists();
+//				super.signOut();
+//			}
+//	}
+//}
