+++ src/test/resources/student/activity/delete-positive.csv
@@ -1,0 +1,3 @@
+enrolmentRecordIndex,activityRecordIndex
+0,1
+0,0
